requirements.txt

# requirements.txt

streamlit>=1.30.0
polars>=0.20.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
